requirements.txt
faiss-cpu>=1.8.0
gradio>=4.44.0
pyspark>=3.5.0
delta-spark>=3.2.0
mlflow>=2.16.0
pypdf>=4.3.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
pytest>=8.3.0
sentence-transformers>=3.0.0
llama-cpp-python>=0.2.90
